move db to a separate file
